# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/": {
      "get": {
        "tags": [
          "Venues"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Venues"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "user_id": {
                  "example": "any"
                },
                "location_name": {
                  "example": "any"
                },
                "city_ward": {
                  "example": "any"
                },
                "prefecture": {
                  "example": "any"
                },
                "phone_num": {
                  "example": "any"
                },
                "address": {
                  "example": "any"
                },
                "venue_email": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "num_seats": {
                  "example": "any"
                },
                "smoking": {
                  "example": "any"
                },
                "outdoor_seating": {
                  "example": "any"
                },
                "venue_url": {
                  "example": "any"
                },
                "photo_link": {
                  "example": "any"
                },
                "type": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user/{user_id}": {
      "get": {
        "tags": [
          "Events"
        ],
        "description": "",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/venue/{venue_id}": {
      "get": {
        "tags": [
          "Events"
        ],
        "description": "",
        "parameters": [
          {
            "name": "venue_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/{event_id}": {
      "patch": {
        "tags": [
          "Events"
        ],
        "description": "",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "edits": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Events"
        ],
        "description": "",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/{user_id}": {
      "get": {
        "tags": [
          "Groups"
        ],
        "description": "",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/{group_id}/privacy": {
      "patch": {
        "tags": [
          "Groups"
        ],
        "description": "",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/{group_id}": {
      "patch": {
        "tags": [
          "Groups"
        ],
        "description": "",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/joingroup": {
      "post": {
        "tags": [
          "Groups"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "group_id": {
                  "example": "any"
                },
                "user_id": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/members/{group_id}": {
      "get": {
        "tags": [
          "Groups"
        ],
        "description": "",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/{email}": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/{email}/activation": {
      "patch": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/{id}": {
      "get": {
        "tags": [
          "Venues"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Venues"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Venues"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/city/{city}": {
      "get": {
        "tags": [
          "Venues"
        ],
        "description": "",
        "parameters": [
          {
            "name": "city",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/prefecture/{prefecture}": {
      "get": {
        "tags": [
          "Venues"
        ],
        "description": "",
        "parameters": [
          {
            "name": "prefecture",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/package/": {
      "post": {
        "tags": [
          "Venues"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "venue_id": {
                  "example": "any"
                },
                "package_name": {
                  "example": "any"
                },
                "package_per_person_cost": {
                  "example": "any"
                },
                "duration": {
                  "example": "any"
                },
                "maximum_number_of_people": {
                  "example": "any"
                },
                "picture_url": {
                  "example": "any"
                },
                "other_notes": {
                  "example": "any"
                },
                "drinks": {
                  "example": "any"
                },
                "food": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/packages/{venue_id}": {
      "get": {
        "tags": [
          "Venues"
        ],
        "description": "",
        "parameters": [
          {
            "name": "venue_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/packages/{package_id}": {
      "delete": {
        "tags": [
          "Venues"
        ],
        "description": "",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Venues"
        ],
        "description": "",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}